Fixe error dataProvider
