+++ admin/src/components/auth/dataProvider.js
@@ -5,7 +5,7 @@ const httpClient = (url, options = {}) => {
   if (!options.headers) {
     options.headers =  new Headers({ Accept: "application/json" });
   }
-  const { token } =  JSON.parse(localStorage.getItem('auth'));
+  const token  =  JSON.parse(localStorage.getItem('auth'));
    options.headers.set("Authorization", `Bearer ${token}`);
   return fetchUtils.fetchJson(url, options);
 };
